# docs/swagger.yaml
basePath: /fab
definitions:
  chaincode.approveChaincodeRequest:
    properties:
      chaincodeid:
        example: test
        type: string
      chaincodepolicy:
        type: string
      chaincodeversion:
        example: v1.0
        type: string
      channelid:
        example: mychanel
        type: string
      isinit:
        type: boolean
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
      packageid:
        type: string
      sequence:
        example: 1
        type: integer
    required:
    - chaincodeid
    - chaincodeversion
    - channelid
    - orgmsps
    - orgnames
    - packageid
    - sequence
    type: object
  chaincode.commitChaincodeRequest:
    properties:
      chaincodeid:
        example: test
        type: string
      chaincodepolicy:
        type: string
      chaincodeversion:
        example: v1.0
        type: string
      channelid:
        example: mychannel
        type: string
      isinit:
        type: boolean
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
      packageid:
        type: string
      sequence:
        example: 1
        type: integer
    required:
    - chaincodeid
    - chaincodeversion
    - channelid
    - orgmsps
    - orgnames
    - packageid
    - sequence
    type: object
  chaincode.initChaincodeRequest:
    properties:
      args:
        example:
        - save
        - a
        - b
        items:
          type: string
        type: array
      chaincodeid:
        example: test
        type: string
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - chaincodeid
    - channelid
    - orgmsps
    - orgnames
    type: object
  chaincode.installChaincodeRequest:
    properties:
      ccpack:
        required:
        - ccpack
        type: object
      chaincodeid:
        example: test
        type: string
      chaincodeversion:
        example: v1.0
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - chaincodeid
    - chaincodeversion
    - orgmsps
    - orgnames
    type: object
  chaincode.invokeChaincodeRequest:
    properties:
      args:
        example:
        - save
        - a
        - b
        items:
          type: string
        type: array
      chaincodeid:
        example: test
        type: string
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - args
    - chaincodeid
    - channelid
    - orgmsps
    - orgnames
    type: object
  chaincode.packageChaincodeRequest:
    properties:
      chaincodeid:
        example: test
        type: string
      chaincodesrcpath:
        example: src/github.com/chaincode
        type: string
      chaincodetype:
        example: GOLANG
        type: string
      chaincodeversion:
        example: v1.0
        type: string
    required:
    - chaincodeid
    - chaincodesrcpath
    - chaincodetype
    - chaincodeversion
    type: object
  chaincode.queryChaincodeRequest:
    properties:
      args:
        example:
        - query
        - a
        items:
          type: string
        type: array
      chaincodeid:
        example: test
        type: string
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - args
    - chaincodeid
    - channelid
    - orgmsps
    - orgnames
    type: object
  chaincode.upgradeChaincodeRequest:
    properties:
      chaincodeid:
        example: test
        type: string
      chaincodepolicy:
        type: string
      chaincodeversion:
        example: v1.0
        type: string
      channelid:
        example: mychanel
        type: string
      isinit:
        type: boolean
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
      packageid:
        type: string
      sequence:
        example: 1
        type: integer
    required:
    - chaincodeid
    - chaincodeversion
    - channelid
    - orgmsps
    - orgnames
    - packageid
    - sequence
    type: object
...
  gintool.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      err_msg:
        description: 内部错误详情
        type: string
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
...
host: localhost:6922
info:
  contact: {}
  description: 描述:fabapi 接口文档.
  license: {}
  title: fabapi API文档
  version: 1.1.0
paths:
  /chaincode/approvecc:
    post:
      consumes:
      - application/json
      description: 功能：批准链码管理
      parameters:
      - description: 批准链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.approveChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 批准链码管理
      tags:
      - 批准链码管理
  /chaincode/commitcc:
    post:
      consumes:
      - application/json
      description: 功能：提交链码管理
      parameters:
      - description: 提交链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.commitChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 提交链码管理
      tags:
      - 提交链码管理
  /chaincode/initcc:
    post:
      consumes:
      - application/json
      description: 功能：初始化链码管理
      parameters:
      - description: 初始化链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.initChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 初始化链码管理
      tags:
      - 初始化链码管理
  /chaincode/installcc:
    post:
      consumes:
      - application/json
      description: 功能：安装链码管理
      parameters:
      - description: 安装链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.installChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 安装链码管理
      tags:
      - 安装链码管理
  /chaincode/invokecc:
    post:
      consumes:
      - application/json
      description: 功能：invoke链码管理
      parameters:
      - description: invoke链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.invokeChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: invoke链码管理
      tags:
      - invoke链码管理
  /chaincode/packagecc:
    post:
      consumes:
      - application/json
      description: 功能：打包链码管理
      parameters:
      - description: 打包链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.packageChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 打包链码管理
      tags:
      - 打包链码管理
  /chaincode/querycc:
    post:
      consumes:
      - application/json
      description: 功能：query链码管理
      parameters:
      - description: query链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.queryChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: query链码管理
      tags:
      - query链码管理
  /chaincode/upgradecc:
    post:
      consumes:
      - application/json
      description: 功能：升级链码管理
      parameters:
      - description: 升级链码管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/chaincode.upgradeChaincodeRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 升级链码管理
      tags:
      - 升级链码管理
...
swagger: "2.0"